finished refactor Calendar component
