+++ src/components/Content/Main/Calendar/Calendar.jsx
@@ -1,137 +1,52 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { ArrowLeftIcon, ArrowRightIcon } from '@heroicons/react/solid';
 import s from "./Calendar.module.css"
 import classNames from "./helper.ts";
 import "./style.css"
-import { monthNames, days } from "../../../../constants/index.js"
 import CalendarMonth from './Calendarmonth';
 import { useRef } from 'react';
-import { taskInfoApi, editTaskApi } from "../../../../api/api.ts"
 import useStore from "../../../../store/useToDoStore.js";
-
+import { monthNames, days } from "../../../../constants/index.js"
+import { isToday, adjustCellHeights, btnClass, eventClass } from "../../../../utils/Calendar_functions.js"
 const Calendar = (props) => {
 
   const openTaskInfoS = useStore((state) => state.openTaskInfoS)
-  const updateTasks = useStore((state) => state.updateTasks)
   const getTaskInfo = useStore((state) => state.getTaskInfo)
   const getTaskStatus = useStore((state) => state.getTaskStatus)
   const tasks = useStore((state) => state.tasks)
   const events = useStore((state) => state.events)
   const handleNewData = useStore((state) => state.handleNewData)
+
+  const month = useStore((state) => state.month)
+  const year = useStore((state) => state.year)
+  
+  const numOfDays = useStore((state) => state.numOfDays)
+  const emptyDays = useStore((state) => state.emptyDays)
+  const getNoOfDays = useStore((state) => state.getNoOfDays)
+
   const nextMonth = useStore((state) => state.nextMonth)
   const prevMonth = useStore((state) => state.prevMonth)
-  const updateMonthYear = useStore((state) => state.updateMonthYear)
   
-  const date = new Date();
-  const [month, setMonth] = useState(date.getMonth());
-  const [year, setYear] = useState(date.getFullYear());
-  const [numOfDays, setNumOfDays] = useState([]);
-  const [emptyDays, setEmptyDays] = useState([]);
+  const goToCurrentMonth = useStore((state) => state.goToCurrentMonth)
+  const handleDrop = useStore((state) => state.handleDrop)
+  
   const cellRefs = useRef([]);
   const draggedItem = useRef(null);
 
-  const isToday = (date) => {
-    const today = new Date();
-    const d = new Date(year, month, date);
-    return today.toDateString() === d.toDateString();
-  };
-
   useEffect(() => {
     if (cellRefs.current) {
-      adjustCellHeights();
+      adjustCellHeights(cellRefs);
     }
   }, [events, month]);
-
-  const adjustCellHeights = () => {
-    cellRefs.current.forEach((cell, index) => {
-      if (cell) {
-        const contentDiv = cell.querySelector('.tasks-container');
-        if (contentDiv) {
-          if (contentDiv.scrollHeight > 0) {
-            cell.style.height = `${32 + (contentDiv.scrollHeight)}px`;
-          } else if (cell.style.height !== '8rem') {
-            cell.style.height = `8rem`;
-          }
-        }
-      }
-    })
-  }
-
-  const getNoOfDays = () => {
-    let i;
-    let daysInMonth = new Date(year, month + 1, 0).getDate();
-    let dayOfWeek = new Date(year, month).getDay();
-    let emptyDaysArray = [];
-    if (dayOfWeek === 0) {
-      for (i = 1; i <= 6; i++) {
-        emptyDaysArray.push(i);
-      }
-    } else {
-      for (i = 1; i <= dayOfWeek - 1; i++) {
-        emptyDaysArray.push(i);
-      }
-    }
-    let daysArray = [];
-    for (i = 1; i <= daysInMonth; i++) {
-      daysArray.push(i);
-    }
-    setEmptyDays(emptyDaysArray);
-    setNumOfDays(daysArray);
-  };
 
   useEffect(() => {
     getNoOfDays();
   }, [month]);
 
+  
   useEffect(() => {
     handleNewData(tasks);
   }, [tasks]);
-
-  const btnClass = (limit) => {
-    return "leading-none rounded-lg transition ease-in-out duration-100 inline-flex cursor-pointer hover:bg-gray-200 p-1 items-center focus:outline-none";
-  };
-
-  const eventClass = (t) => {
-    const hexToRgb = (hex) => {
-      const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-      return result
-        ? {
-          r: parseInt(result[1], 16),
-          g: parseInt(result[2], 16),
-          b: parseInt(result[3], 16),
-        }
-        : null;
-    };
-
-    const rgb = hexToRgb(t);
-
-    if (!rgb) {
-      return {
-        borderColor: '#fff',
-        color: '#000',
-        backgroundColor: '#fff',
-        textAlign: 'center',
-      };
-    }
-
-    const { r, g, b } = rgb;
-    const isLightBackground = (r * 0.299 + g * 0.587 + b * 0.114) > 186;
-    const textColor = isLightBackground ? '#000' : '#fff';
-
-    return {
-      borderColor: t,
-      color: textColor,
-      backgroundColor: `rgb(${r},${g},${b},1)`,
-      textAlign: 'center',
-    }
-
-  };
-
-  const goToCurrentMonth = () => {
-    const currentDate = new Date();
-    setMonth(currentDate.getMonth());
-    setYear(currentDate.getFullYear());
-  };
 
   const handleDragStart = (e, event) => {
     draggedItem.current = event.task_id;
@@ -152,41 +67,19 @@ const Calendar = (props) => {
       e.target.closest(`.${s.adaptive}`).classList.remove(s.dragover)
     }
   };
-  const handleDrop = async (e, date) => {
-    if (e.target.closest(`.${s.adaptive}`)) {
-      e.target.closest(`.${s.adaptive}`).classList.remove(s.dragover)
-    }
-    e.preventDefault();
-    const taskId = e.dataTransfer.getData("text/plain");
-    if (draggedItem.current && taskId) {
-      try {
-        const response = await taskInfoApi(taskId)
-        const newDate = new Date(year, month, date + 1)
-        const taskData = {
-          ...response,
-          date: newDate.toISOString().slice(0, 10)
-        };
-        await editTaskApi(taskId, taskData)
-        updateTasks()
-      } catch (error) {
-        console.error('Ошибка при загрузке задачи для перетаскивания:', error);
-      }
-      draggedItem.current = null
-    }
-  };
-
+  
   return (
     <>
       <div className="container mx-auto py-4 px-6">
         <div className="bg-white rounded-lg shadow overflow-hidden">
-          <div className= {["flex items-center justify-between px-6 py-4", s.container_adaptive].join(" ")}>
+          <div className={["flex items-center justify-between px-6 py-4", s.container_adaptive].join(" ")}>
             <div className={s.content}>
               <div className={s.content_flex}>
                 <span className="text-lg font-bold text-gray-800">
                   {monthNames[month]}
-                  <CalendarMonth updateMonthAndYear={updateMonthYear} year={year} />
+                  <CalendarMonth/>
                 </span>
-                <span className={["ml-1 text-lg text-gray-600 font-normal",s.year].join(" ")}>
+                <span className={["ml-1 text-lg text-gray-600 font-normal", s.year].join(" ")}>
                   {year}
                 </span>
               </div>
@@ -194,7 +87,7 @@ const Calendar = (props) => {
             </div>
             <div className={s.click}>
               <button onClick={openTaskInfoS} className={s.addtask}>Добавить задачу</button>
-              <div className={["border rounded-lg px-1 pt-1",s.arrow].join(" ")}>
+              <div className={["border rounded-lg px-1 pt-1", s.arrow].join(" ")}>
                 <button
                   type="button"
                   onClick={() => prevMonth()}
@@ -221,7 +114,7 @@ const Calendar = (props) => {
               style={{ marginBottom: "-37px" }}>
               {days.map((day) => (
                 <div key={day} className={"px-2 py-2 w-[14.28%]"}>
-                  <div  className="text-gray-600 text-sm uppercase tracking-wide font-bold text-center">
+                  <div className="text-gray-600 text-sm uppercase tracking-wide font-bold text-center">
                     {day}
                   </div>
                 </div>
@@ -238,7 +131,7 @@ const Calendar = (props) => {
                 <div
                   ref={el => cellRefs.current[index] = el}
                   onDragOver={handleDragOver}
-                  onDrop={(e) => handleDrop(e, date)}
+                  onDrop={(e) => handleDrop(e, date, draggedItem, s)}
                   onDragEnter={handleDragEnter}
                   onDragLeave={handleDragLeave}
                   onClick={openTaskInfoS}
@@ -249,7 +142,7 @@ const Calendar = (props) => {
                   <div
                     style={{ marginTop: "20px" }}
                     className={classNames(
-                      isToday(date)
+                      isToday(year, month, date)
                         ? "bg-blue-500 text-white"
                         : "text-gray-700 hover:bg-blue-200", s.day_container,
                       "inline-flex w-6 h-6 items-center justify-center cursor-pointer text-center leading-none rounded-full transition ease-in-out duration-100"
