--- src/utils/utils.js
@@ -1,16 +1,0 @@
-
-export const openTaskInfo = (e) => {
-  let dateString = e?.currentTarget?.getAttribute('data-date');
-  if (!dateString) {
-    const today = new Date();
-    const year = today.getFullYear();
-    const month = String(today.getMonth() + 1).padStart(2, '0');
-    const day = String(today.getDate()).padStart(2, '0');
-    dateString = `${year}-${month}-${day}`;
-  }
-  const dateObj = new Date(dateString);
-  const year = dateObj.getFullYear();
-  const month = String(dateObj.getMonth() + 1).padStart(2, '0');
-  const day = String(dateObj.getDate()).padStart(2, '0');
-  const formattedDate = `${year}-${month}-${day}`;
-};
